Updating to include token owner in method calls
There are modifiers which check the message sender. If calling the functions via a command line the contract owner/message sender is not implied. Obviously using MetaMask etc. does this behind the scenes. Below are a couple of examples of how an error will occur in the event that a message sender is not explicitly added to the method call in the command line.
```
deployedContract.acceptOwnership();
Error: invalid address
    at web3-cmt.js:12250:13
    at web3-cmt.js:12080:18
    at web3-cmt.js:13332:28
    at map (<native code>)
    at web3-cmt.js:13331:12
    at web3-cmt.js:13357:18
    at web3-cmt.js:13382:23
    at web3-cmt.js:12448:16
    at apply (<native code>)
    at web3-cmt.js:12534:16
```

```
deployedContract.issue({from:"0xf7fca0066bde97e5509591afc9f669991e03c25c"},10000000000);
Error: Invalid number of arguments to Solidity function
    at web3-cmt.js:11459:20
    at web3-cmt.js:12352:15
    at web3-cmt.js:12368:5
    at web3-cmt.js:12441:19
    at apply (<native code>)
    at web3-cmt.js:12534:16
    at <anonymous>:1:1
```